config: Support coalesced explicit and SPM flushes
The current implementation assumes that a flush is triggered either by
a write to `cfg_flush` *or* to `cfg_spm`, but never to both at the same
time. In particular, when reconfiguring a way from SPM to LLC while any
bit of `cfg_flush` is set, this can cause this way to remain marked as
`flushed` and functioning neither as SPM nor LLC. In our scenario, the
LLC was still in bypass mode even though some ways have been configured
as LLC.

Fix this by reducing the if/else block to an invariant.

Signed-off-by: Nils Wistoff <nwistoff@iis.ee.ethz.ch>